+++ src/pages/Main.css
@@ -1,0 +1,58 @@
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  }
+  
+  body {
+    background-color: #DCA876;
+    font-family: 'Arial', sans-serif;
+  }
+  
+  .app {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+    width: 100%;
+    overflow: hidden;
+  }
+  
+  .tamagotchi-container {
+    position: relative;
+    width: 100%;
+    max-width: 390px;
+    height: 100vh;
+    max-height: 844px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .top-panel-wrapper {
+    position: absolute;
+    top: 30px;
+    width: 90%;
+    z-index: 3;
+  }
+  
+  .action-buttons-wrapper {
+    position: absolute;
+    bottom: 65px;
+    width: 95%;
+    z-index: 3;
+  }
+  
+  @media (max-height: 700px) {
+    .action-buttons-wrapper {
+      bottom: 45px;
+    }
+  }
+  
+  @media (max-width: 375px) {
+    .action-buttons-wrapper {
+      width: 100%;
+    }
+  }
+  
+
